requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.10.0
bootstrap4==0.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.6
django-allauth==0.46.0
django-bootstrap4==21.1
django-elasticsearch-dsl==7.2.1
django-embed-video==1.4.0
elasticsearch==7.15.2
elasticsearch-dsl==7.4.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
mysqlclient==2.0.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-logstash==0.4.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
